// src/app/core/model/database/rolProyecto.model.ts
export class RolProyectoEntity {
    nombre:             string          = '';
    descripcion:        string          = '';
    hora:               number          = 0;
    id_PROYECTO:        number | null   = null;
    id_ROL_NEGOCIO:     number | null   = null;
    id_ROL_PROYECTO:    number | null   = null;
}